+++ artiq/gateware/rtio/core.py
@@ -290,6 +290,7 @@ class Core(Module, AutoCSR):
         self.cri = cri.Interface()
         self.reset = CSR()
         self.reset_phy = CSR()
+        self.async_error = CSR(2)
         self.comb += self.cri.arb_gnt.eq(1)
 
         # Clocking/Reset
@@ -321,11 +322,31 @@ class Core(Module, AutoCSR):
         # Managers
         self.submodules.counter = RTIOCounter(len(self.cri.timestamp) - fine_ts_width)
 
-        i_datas, i_timestamps = [], []
+        # Collision is not an asynchronous error with local RTIO, but
+        # we treat it as such for consistency with DRTIO, where collisions
+        # are reported by the satellites.
+        o_underflow = Signal()
+        o_sequence_error = Signal()
+        o_collision = Signal()
+        o_busy = Signal()
+        self.sync.rsys += [
+            If(self.cri.cmd == cri.commands["write"],
+                o_underflow.eq(0),
+                o_sequence_error.eq(0),
+            ),
+            If(self.async_error.re,
+                If(self.async_error.r[0], o_collision.eq(0)),
+                If(self.async_error.r[1], o_busy.eq(0)),
+            )
+        ]
+
         o_statuses, i_statuses = [], []
+        i_datas, i_timestamps = [], []
+        i_ack = Signal()
         sel = self.cri.chan_sel[:16]
         for n, channel in enumerate(channels):
             if isinstance(channel, LogChannel):
+                o_statuses.append(1)
                 i_datas.append(0)
                 i_timestamps.append(0)
                 i_statuses.append(0)
@@ -347,29 +368,13 @@ class Core(Module, AutoCSR):
 
             self.comb += o_manager.we.eq(selected & (self.cri.cmd == cri.commands["write"]))
 
-            underflow = Signal()
-            sequence_error = Signal()
-            collision = Signal()
-            busy = Signal()
             self.sync.rsys += [
-                If(self.cri.cmd == cri.commands["o_underflow_reset"],
-                   underflow.eq(0)),
-                If(self.cri.cmd == cri.commands["o_sequence_error_reset"],
-                   sequence_error.eq(0)),
-                If(self.cri.cmd == cri.commands["o_collision_reset"],
-                   collision.eq(0)),
-                If(self.cri.cmd == cri.commands["o_busy_reset"],
-                   busy.eq(0)),
-                If(o_manager.underflow, underflow.eq(1)),
-                If(o_manager.sequence_error, sequence_error.eq(1)),
-                If(o_manager.collision, collision.eq(1)),
-                If(o_manager.busy, busy.eq(1))
+                If(o_manager.underflow, o_underflow.eq(1)),
+                If(o_manager.sequence_error, o_sequence_error.eq(1)),
+                If(o_manager.collision, o_collision.eq(1)),
+                If(o_manager.busy, o_busy.eq(1))
             ]
-            o_statuses.append(Cat(~o_manager.writable,
-                                  underflow,
-                                  sequence_error,
-                                  collision,
-                                  busy))
+            o_statuses.append(o_manager.writable)
 
             if channel.interface.i is not None:
                 i_manager = _InputManager(channel.interface.i, self.counter,
@@ -386,42 +391,49 @@ class Core(Module, AutoCSR):
                 else:
                     i_timestamps.append(0)
 
-                self.comb += i_manager.re.eq(selected & (self.cri.cmd == cri.commands["read"]))
-
                 overflow = Signal()
                 self.sync.rsys += [
-                    If(selected & (self.cri.cmd == cri.commands["i_overflow_reset"]),
+                    If(selected & i_ack,
                        overflow.eq(0)),
                     If(i_manager.overflow,
                        overflow.eq(1))
                 ]
-                i_statuses.append(Cat(i_manager.readable, overflow))
+                self.comb += i_manager.re.eq(selected & i_ack & ~overflow)
+                i_statuses.append(Cat(i_manager.readable & ~overflow, overflow))
 
             else:
                 i_datas.append(0)
                 i_timestamps.append(0)
                 i_statuses.append(0)
 
-        i_status_raw = Signal(2)
-        self.sync.rsys += i_status_raw.eq(Array(i_statuses)[sel])
+        o_status_raw = Signal()
+        self.comb += [
+            o_status_raw.eq(Array(o_statuses)[sel]),
+            self.cri.o_status.eq(Cat(
+                ~o_status_raw, o_underflow, o_sequence_error)),
+            self.async_error.w.eq(Cat(o_collision, o_busy))
+        ]
 
+        i_status_raw = Signal(2)
+        self.comb += i_status_raw.eq(Array(i_statuses)[sel])
         input_timeout = Signal.like(self.cri.timestamp)
         input_pending = Signal()
         self.sync.rsys += [
+            i_ack.eq(0),
+            If(i_ack,
+                self.cri.i_status.eq(Cat(~i_status_raw[0], i_status_raw[1], 0)),
+                self.cri.i_data.eq(Array(i_datas)[sel]),
+                self.cri.i_timestamp.eq(Array(i_timestamps)[sel]),
+            ),
             If((self.cri.counter >= input_timeout) | (i_status_raw != 0),
+                If(input_pending, i_ack.eq(1)),
                 input_pending.eq(0)
             ),
-            If(self.cri.cmd == cri.commands["read_request"],
+            If(self.cri.cmd == cri.commands["read"],
                 input_timeout.eq(self.cri.timestamp),
-                input_pending.eq(1)
+                input_pending.eq(1),
+                self.cri.i_status.eq(0b100)
             )
         ]
 
-        self.comb += [
-            self.cri.i_data.eq(Array(i_datas)[sel]),
-            self.cri.i_timestamp.eq(Array(i_timestamps)[sel]),
-            self.cri.o_status.eq(Array(o_statuses)[sel]),
-            self.cri.i_status.eq(Cat(~i_status_raw[0], i_status_raw[1], input_pending)),
-            self.cri.counter.eq(self.counter.value_sys << fine_ts_width)
-        ]
-
+        self.comb += self.cri.counter.eq(self.counter.value_sys << fine_ts_width)